// src/app/model/persona.model.ts
export class PersonaModel {
    idpersona: any;
    nombre: string;
    apellido: string;
    dni: string;
    cargo: string;
}
export class UsuariModel {
    idusuaio?:any;
    username:string;
    password: string;
    rol:any;
}
export class RolModel {
    idrol: any;
    descripcionrol: string;
}
export class SelectModel {
    value: any;
    label: string;
}